requirements.txt
openai==1.33.0
requests==2.32.3
pytz==2024.1
discord.py==2.3.2
ipython==8.25.0
google-generativeai==0.6.0
pillow==10.3.0